Split off the debug menu entries from the debug menu module
